requirements.txt
tqdm==4.63.1
torch==1.12.1+cu113
seaborn==0.11.2
matplotlib==3.3.4
Pillow==8.4.0
pytorch_lightning==1.9.0
transformers==4.18.0
pycocotools
torchmetrics==1.4.0.post0
nltk==3.7
scikit_learn==0.24.2
wordcloud==1.8.1
torchvision==0.13.1+cu113
requests==2.27.1
tabulate==0.8.9
packaging==21.3
pandas==1.1.5
dgl_cu111==0.6.1
Cython==0.29.28
numpy==1.21.0
FactualSceneGraph
sentence_transformers
SPARQLWrapper==2.0.0
jieba==0.42.1
spacy==3.8.2
